requirements.txt
annotated-types==0.6.0
anyio==4.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==2.0.12
dnspython==1.16.0
feedparser==6.0.10
gnews==0.3.6
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==3.12.3
pyTelegramBotAPI==4.14.0
python-dotenv==0.19.2
python-telegram-bot==20.6
requests==2.26.0
sgmllib3k==1.0.0
sniffio==1.3.0
soupsieve==2.5
typing_extensions==4.8.0
urllib3==1.26.18
